+++ pkg/analyzer/lib/src/dart/element/type_system.dart
@@ -539,7 +539,7 @@ class TypeSystemImpl implements TypeSystem {
   FunctionTypeImpl? getCallMethodType(DartType t) {
     if (t is InterfaceTypeImpl) {
       return t
-          .lookUpMethod3(MethodElement.CALL_METHOD_NAME, t.element3.library2)
+          .lookUpMethod(MethodElement.CALL_METHOD_NAME, t.element3.library2)
           ?.type;
     }
     return null;